Finally Got the Score Working
Apparently the GUI Text is different in the version of Unity I am using.
I restarted the assignment, used an updated guide and was able to
resolve the issue. Also, I didn't like how the score was just a number
represented as a String. What I did was place an in variable into the
script and concatenated that with the String "Score: " to make the
display look cleaner. I'm saving now because my battery is getting low
and I don't want to lose my work.