Finished!
Man, it feels so good to be done. This game was slightly more difficult
than the first one we did. I had a lot of issues completing it, but I
stayed persistent. As you can tell by the time, I am late again working
on these assignments, but there is a big satisfaction from getting them
to work.